+++ control_system/smooth_change_speed.h
@@ -9,7 +9,7 @@
 #define	SMOOTH_CHANGE_SPEED_H
 
 void smooth_change_current_speed_init();
-void smooth_change_current_speed(uint16_t* currentSpeed, uint32_t nowPulses, uint32_t needPulses);
+void smooth_change_current_speed(uint8_t* currentSpeed, uint32_t nowPulses, uint32_t needPulses);
 
 #endif	/* SMOOTH_CHANGE_SPEED_H */
